# docs/swagger.yaml
basePath: /
definitions:
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handler.MessageResponse:
    properties:
      message:
        type: string
    type: object
  models.Account:
    properties:
      account_number:
        type: string
      balance:
        type: number
      created_at:
        type: string
      owner:
        type: string
      updated_at:
        type: string
    type: object
  models.CreateAccountDTO:
    properties:
      balance:
        type: number
      owner:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample implementation of a RESTful API for a bank account.
  title: Bank Account API
  version: "1.0"
paths:
  /account:
    post:
      consumes:
      - application/json
      description: Create a new account with the input payload
      parameters:
      - description: Account
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/models.CreateAccountDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create a new account
      tags:
      - account
  /account/{accountNumber}:
    delete:
      description: Delete an account by account number
      parameters:
      - description: Account Number
        in: path
        name: accountNumber
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete an account by account number
      tags:
      - account
    get:
      description: Get an account by account number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get an account by account number
      tags:
      - account
    put:
      consumes:
      - application/json
      description: Update an account by account number
      parameters:
      - description: Account
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateAccountDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update an account by account number
      tags:
      - account
  /accounts:
    get:
      description: Get all accounts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Account'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get all accounts
      tags:
      - accounts
swagger: "2.0"